Steuerung etwas angepasst, sodass man ein wenig rumfliegen kann ;)

+++ Asteroids/Asteroids/Components/Motion.h
@@ -2,6 +2,7 @@
 #define _USE_MATH_DEFINES
 #include "Unit.h"
 #include <cmath>
+#include <iostream>
 
 class Motion : public Unit
 {
@@ -12,7 +13,7 @@ public:
 	sf::Packet* getPacket(int clientId, int id);
 	void refresh(sf::Packet packet);
 
-	void increaseSpeed(int rotation, int accelerationRate);
+	void increaseSpeed(int rotation, float accelerationRate);
 
 	float getSpeedX();
 	float getSpeedY();
